requirements.txt
fastapi
uvicorn[standard]
pytesseract
Pillow
python-multipart
openai